# YJIT: Propagate Array, Hash, and String classes